Address review: merge capture scripts into one file, trim near-duplicate tests
Per review feedback on PR #7054:

- Merges capture_common.py, capture_hero.py, capture_showcase.py, and
  capture_streaming.py into a single tools/docs/media/visualizers/
  capture_visualizer.py, invoked as `capture_visualizer.py [hero|showcase|
  streaming|all]` via argparse subparsers (default: all three in sequence).
  Resolves every name collision this required (configure_playback_*, _main_*,
  per-pipeline _make_*_visualizer_cfg/_headless/window-size constants/task-
  config dicts) and updates every gym entry-point string, --external_callback
  string, and self-invocation subprocess path accordingly. Live-verified all
  three pipelines end to end against the real GPU capture path (calibration,
  play.py --external_callback resolution, window capture, ffmpeg
  post-processing all confirmed working; showcase produced a full real clip).

- Combines test_rerun_viser_camera.py's 4 near-identical tests into 2 (one
  per backend), covering the same scenarios via sequential assertions
  instead of one test per behavior.
